fix: Resolve supervisor and coordinator permission issues on production
- Fix strict type comparisons (string vs int) in Policies and Controllers
- Fix 'Your Trainee' badge display logic in supervisor view
- Fix Acceptance Letter download permissions
- Fix Final and Monthly Evaluation creation/viewing permissions
- Remove temporary deployment guides

+++ app/Http/Controllers/SupervisorAcceptanceController.php
@@ -156,7 +156,7 @@ class SupervisorAcceptanceController extends Controller
         // Check if student already has a different supervisor
         if ($student->studentProfile && $student->studentProfile->supervisor_id) {
             // If it's a different supervisor trying to view, block them
-            if (Auth::id() !== $student->studentProfile->supervisor_id) {
+            if ((int) Auth::id() !== (int) $student->studentProfile->supervisor_id) {
                 return redirect()->route('supervisor.students.search')
                     ->with('error', 'This student already has a supervisor. You cannot accept this student.');
             }
@@ -181,7 +181,7 @@ class SupervisorAcceptanceController extends Controller
             ->firstOrFail();
 
         // Check if student already has a supervisor
-        if ($student->studentProfile && $student->studentProfile->supervisor_id && $student->studentProfile->supervisor_id !== Auth::id()) {
+        if ($student->studentProfile && $student->studentProfile->supervisor_id && (int) $student->studentProfile->supervisor_id !== (int) Auth::id()) {
             return redirect()->route('supervisor.students.search')
                 ->with('error', 'This student already has a supervisor.');
         }
@@ -203,7 +203,7 @@ class SupervisorAcceptanceController extends Controller
             ->firstOrFail();
 
         // Check if student already has a supervisor
-        if ($student->studentProfile && $student->studentProfile->supervisor_id && $student->studentProfile->supervisor_id !== Auth::id()) {
+        if ($student->studentProfile && $student->studentProfile->supervisor_id && (int) $student->studentProfile->supervisor_id !== (int) Auth::id()) {
             return back()->with('error', 'This student already has a supervisor.');
         }
 
@@ -573,7 +573,7 @@ class SupervisorAcceptanceController extends Controller
         $letter = AcceptanceLetter::with('student', 'company')->findOrFail($letterId);
 
         // Ensure the logged-in supervisor owns this letter
-        if ($letter->supervisor_user_id !== Auth::id()) {
+        if ((int) $letter->supervisor_user_id !== (int) Auth::id()) {
             abort(403);
         }
